Refactor identity management
Refactor IdentityManager to expose credentials as stateflows and
deprecated redundant methods. The already existing methods needed to be
constantly called to get the updated values which was not safe in case
user has not performed onboarding and also makes it harder to support
username. Changing it to expose a stateflow makes it safer and more
efficient